support `config[:timeout]` for `Faraday` request timeout
to enable configuring the ClickHouse HTTP API client request timeout,
this adds a check against the `@config` hash passed to the initializer
of the `Clickhouse::Client` class for a `timeout` key while creating the
underlying `Faraday` HTTP client instance. specifically, this sets the
default timeout for requests dispatched from that instance. this also
introduces an explicit default timeout of 60 seconds.